feat：蓝盾本地开发实现 #9576
