+++ src/backend/ci/core/buildless/biz-buildless/src/main/kotlin/com/tencent/devops/buildless/rejected/JumpPolicy.kt
@@ -29,19 +29,19 @@ package com.tencent.devops.buildless.rejected
 
 import com.tencent.devops.buildless.pojo.BuildLessStartInfo
 import com.tencent.devops.buildless.pojo.BuildLessTask
-import com.tencent.devops.buildless.utils.RedisUtils
+import com.tencent.devops.buildless.utils.BuildlessRedisUtils
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 
 @Service
 class JumpPolicy @Autowired constructor(
-    private val redisUtils: RedisUtils
+    private val buildlessRedisUtils: BuildlessRedisUtils
 ) : RejectedExecutionHandler {
     override fun rejectedExecution(buildLessStartInfo: BuildLessStartInfo): Boolean {
         with(buildLessStartInfo) {
             logger.info("$buildId|$vmSeqId|$executionCount jump allocate container.")
-            redisUtils.rightPushBuildLessReadyTask(
+            buildlessRedisUtils.rightPushBuildLessReadyTask(
                 BuildLessTask(
                     projectId = projectId,
                     pipelineId = pipelineId,
