+++ src/backend/ci/core/buildless/biz-buildless/src/main/kotlin/com/tencent/devops/buildless/schedule/BuildlessSystemInfoRunner.kt
@@ -32,7 +32,7 @@ import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 
 @Component
-class SystemInfoRunner {
+class BuildlessSystemInfoRunner {
 
     @Scheduled(cron = "0/5 * * * * ?")
     fun startCollect() {
