+++ src/backend/ci/core/buildless/biz-buildless/src/main/kotlin/com/tencent/devops/buildless/utils/BuildlessRedisUtils.kt
@@ -36,7 +36,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
 
 @Component
-class RedisUtils @Autowired constructor(
+class BuildlessRedisUtils @Autowired constructor(
     private val redisOperation: RedisOperation,
     private val objectMapper: ObjectMapper
 ) {
@@ -139,6 +139,6 @@ class RedisUtils @Autowired constructor(
     }
 
     companion object {
-        private val logger = LoggerFactory.getLogger(RedisUtils::class.java)
+        private val logger = LoggerFactory.getLogger(BuildlessRedisUtils::class.java)
     }
 }
