+++ src/backend/ci/core/dispatch-docker/api-dispatch-docker/src/main/kotlin/com/tencent/devops/dispatch/docker/api/builds/DispatchBuildDockerHostResource.kt
@@ -47,7 +47,7 @@ import javax.ws.rs.core.MediaType
 @Path("/dockerhost")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)@Suppress("ALL")
-interface BuildDockerHostResource {
+interface DispatchBuildDockerHostResource {
     @GET
     @Path("/resource-config/pipelines/{pipelineId}/vmSeqs/{vmSeqId}")
     @ApiOperation("获取蓝盾项目的docker性能配置")
