+++ src/backend/ci/core/dispatch-docker/api-dispatch-docker/src/main/kotlin/com/tencent/devops/dispatch/docker/api/service/ServiceDispatchDockerHostResource.kt
@@ -54,7 +54,7 @@ import javax.ws.rs.core.MediaType
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 @ServiceInterface("dispatch-docker") // 指明接入到哪个微服务
-interface ServiceDockerHostResource {
+interface ServiceDispatchDockerHostResource {
 
     @ApiOperation("获取dockerhost列表")
     @GET
