+++ src/backend/ci/core/dispatch-docker/biz-dispatch-docker/src/main/kotlin/com/tencent/devops/dispatch/docker/config/DispatchDockerScheduleConfig.kt
@@ -1,4 +1,4 @@
-package com.tencent.devops.buildless.config
+package com.tencent.devops.dispatch.docker.config
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.scheduling.annotation.SchedulingConfigurer
@@ -8,7 +8,7 @@ import java.util.concurrent.ThreadPoolExecutor
 import java.util.concurrent.TimeUnit
 
 @Configuration
-class ScheduleConfig : SchedulingConfigurer {
+class DispatchDockerScheduleConfig : SchedulingConfigurer {
     override fun configureTasks(taskRegistrar: ScheduledTaskRegistrar) {
         val scheduler = ScheduledThreadPoolExecutor(5)
         scheduler.maximumPoolSize = 5
