+++ src/backend/ci/core/dispatch-docker/biz-dispatch-docker/src/main/kotlin/com/tencent/devops/dispatch/docker/controller/ServiceDispatchDockerHostResourceImpl.kt
@@ -31,22 +31,22 @@ import com.tencent.devops.common.api.exception.ParamBlankException
 import com.tencent.devops.common.api.pojo.Page
 import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.web.RestResource
-import com.tencent.devops.dispatch.docker.api.service.ServiceDockerHostResource
+import com.tencent.devops.dispatch.docker.api.service.ServiceDispatchDockerHostResource
 import com.tencent.devops.dispatch.docker.pojo.DockerHostZone
 import com.tencent.devops.dispatch.docker.pojo.DockerIpInfoVO
 import com.tencent.devops.dispatch.docker.pojo.SpecialDockerHostVO
 import com.tencent.devops.dispatch.docker.service.DispatchDockerService
-import com.tencent.devops.dispatch.docker.service.DockerHostBuildService
+import com.tencent.devops.dispatch.docker.service.DispatchDockerHostBuildService
 import com.tencent.devops.dispatch.docker.service.DockerHostZoneTaskService
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 
 @RestResource@Suppress("ALL")
-class ServiceDockerHostResourceImpl @Autowired constructor(
-    private val dockerHostBuildService: DockerHostBuildService,
+class ServiceDispatchDockerHostResourceImpl @Autowired constructor(
+    private val dispatchDockerHostBuildService: DispatchDockerHostBuildService,
     private val dispatchDockerService: DispatchDockerService,
     private val dockerHostZoneTaskService: DockerHostZoneTaskService
-) : ServiceDockerHostResource {
+) : ServiceDispatchDockerHostResource {
     override fun list(page: Int?, pageSize: Int?): Page<DockerHostZone> {
         checkParams(page, pageSize)
         val realPage = page ?: 1
@@ -62,7 +62,7 @@ class ServiceDockerHostResourceImpl @Autowired constructor(
     }
 
     override fun updateContainerId(buildId: String, vmSeqId: Int, containerId: String): Result<Boolean> {
-        dockerHostBuildService.updateContainerId(buildId, vmSeqId, containerId)
+        dispatchDockerHostBuildService.updateContainerId(buildId, vmSeqId, containerId)
         return Result(true)
     }
 
@@ -78,7 +78,7 @@ class ServiceDockerHostResourceImpl @Autowired constructor(
     }
 
     companion object {
-        private val logger = LoggerFactory.getLogger(ServiceDockerHostResourceImpl::class.java)
+        private val logger = LoggerFactory.getLogger(ServiceDispatchDockerHostResourceImpl::class.java)
     }
 
     fun checkParams(page: Int?, pageSize: Int?) {
