+++ src/backend/ci/core/dispatch-docker/biz-dispatch-docker/src/main/kotlin/com/tencent/devops/dispatch/docker/service/DispatchDockerHostBuildService.kt
@@ -55,7 +55,7 @@ import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Service
 
 @Service
-class DockerHostBuildService @Autowired constructor(
+class DispatchDockerHostBuildService @Autowired constructor(
     private val dslContext: DSLContext,
     private val dockerHostClient: DockerHostClient,
     private val pipelineDockerEnableDao: PipelineDockerEnableDao,
@@ -349,6 +349,6 @@ class DockerHostBuildService @Autowired constructor(
     }
 
     companion object {
-        private val LOG = LoggerFactory.getLogger(DockerHostBuildService::class.java)
+        private val LOG = LoggerFactory.getLogger(DispatchDockerHostBuildService::class.java)
     }
 }
