+++ src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/engine/service/rule/processor/BkVarProcessorService.kt
@@ -33,10 +33,10 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 
 @Service("BkVarProcessor")
-class BkVarProcessorService constructor(
-    private val dslContext: DSLContext
-) : ProcessorService {
+class BkVarProcessorService : ProcessorService {
 
+    @Autowired
+    private lateinit var dslContext: DSLContext
 
     @Autowired
     private lateinit var pipelineBuildVarDao: PipelineBuildVarDao
