+++ src/backend/ci/core/repository/biz-repository/src/main/kotlin/com/tencent/devops/repository/resources/ServiceRepositoryGitCheckResourceImpl.kt
@@ -5,12 +5,12 @@ import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.web.RestResource
 import com.tencent.devops.repository.api.ServiceRepositoryGitCheckResource
 import com.tencent.devops.repository.pojo.RepositoryGitCheck
-import com.tencent.devops.repository.service.GitCheckService
+import com.tencent.devops.repository.service.RepoGitCheckService
 import org.springframework.beans.factory.annotation.Autowired
 
 @RestResource
 class ServiceRepositoryGitCheckResourceImpl @Autowired constructor(
-    private val gitCheckService: GitCheckService
+    private val repoGitCheckService: RepoGitCheckService
 ) : ServiceRepositoryGitCheckResource {
     override fun getGitCheck(
         pipelineId: String,
@@ -18,14 +18,14 @@ class ServiceRepositoryGitCheckResourceImpl @Autowired constructor(
         context: String,
         repositoryConfig: RepositoryConfig
     ): Result<RepositoryGitCheck?> {
-        return Result(gitCheckService.getGitCheck(pipelineId, repositoryConfig, commitId, context))
+        return Result(repoGitCheckService.getGitCheck(pipelineId, repositoryConfig, commitId, context))
     }
 
     override fun createGitCheck(gitCheck: RepositoryGitCheck) {
-        gitCheckService.creatGitCheck(gitCheck)
+        repoGitCheckService.creatGitCheck(gitCheck)
     }
 
     override fun updateGitCheck(gitCheckId: Long, buildNumber: Int) {
-        gitCheckService.updateGitCheck(gitCheckId = gitCheckId, buildNumber = buildNumber)
+        repoGitCheckService.updateGitCheck(gitCheckId = gitCheckId, buildNumber = buildNumber)
     }
 }
