+++ src/backend/ci/core/stream/api-stream/src/main/kotlin/com/tencent/devops/stream/api/external/StreamExternalStreamResource.kt
@@ -27,51 +27,39 @@
 
 package com.tencent.devops.stream.api.external
 
-import com.tencent.devops.common.api.pojo.Result
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
 import javax.ws.rs.Consumes
 import javax.ws.rs.GET
-import javax.ws.rs.HeaderParam
-import javax.ws.rs.POST
 import javax.ws.rs.Path
+import javax.ws.rs.PathParam
 import javax.ws.rs.Produces
 import javax.ws.rs.QueryParam
 import javax.ws.rs.core.MediaType
-import javax.ws.rs.core.Response
 
-@Api(tags = ["EXTERNAL_GITHUB"], description = "External-Github")
-@Path("/external/github/")
+@Api(tags = ["EXTERNAL_STREAM"], description = "外部-STREAM资源获取")
+@Path("/external/stream")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-interface ExternalGithubResource {
+interface StreamExternalStreamResource {
 
-    @ApiOperation("Github仓库提交")
-    @POST
-    @Path("/webhook/commit")
-    fun webhookCommit(
-        @ApiParam(value = "事件类型", required = true)
-        @HeaderParam("X-GitHub-Event")
-        event: String,
-        @ApiParam(value = "事件ID", required = true)
-        @HeaderParam("X-Github-Delivery")
-        guid: String,
-        @ApiParam(value = "secretKey签名(sha1)", required = true)
-        @HeaderParam("X-Hub-Signature")
-        signature: String,
-        body: String
-    ): Result<Boolean>
-
-    @ApiOperation("Github Oauth回调")
+    @ApiOperation("获取流水线徽章")
+    @Produces("image/svg+xml")
     @GET
-    @Path("/oauth/callback")
-    fun oauthCallback(
-        @ApiParam(value = "code")
-        @QueryParam("code")
-        code: String,
-        @ApiParam(value = "state")
-        @QueryParam("state")
-        state: String?
-    ): Response
+    @Path("/projects/{gitProjectId}/pipelines/badge")
+    fun getPipelineBadge(
+        @ApiParam("Git仓库ID", required = true)
+        @PathParam("gitProjectId")
+        gitProjectId: Long,
+        @ApiParam("流水线文件名称", required = true)
+        @QueryParam("file_path")
+        filePath: String,
+        @ApiParam("分支名称", required = false)
+        @QueryParam("branch")
+        branch: String?,
+        @ApiParam("触发方式", required = false)
+        @QueryParam("object_kind")
+        objectKind: String?
+    ): String
 }
