+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/resources/external/StreamExternalScmResourceImpl.kt
@@ -35,7 +35,7 @@ import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.api.util.JsonUtil
 import com.tencent.devops.common.web.RestResource
 import com.tencent.devops.common.webhook.pojo.code.git.GitReviewEvent
-import com.tencent.devops.stream.api.external.ExternalScmResource
+import com.tencent.devops.stream.api.external.StreamExternalScmResource
 import com.tencent.devops.stream.trigger.mq.streamRequest.StreamRequestDispatcher
 import com.tencent.devops.stream.trigger.mq.streamRequest.StreamRequestEvent
 import org.slf4j.LoggerFactory
@@ -43,13 +43,13 @@ import org.springframework.amqp.rabbit.core.RabbitTemplate
 import org.springframework.beans.factory.annotation.Autowired
 
 @RestResource
-class ExternalScmResourceImpl @Autowired constructor(
+class StreamExternalScmResourceImpl @Autowired constructor(
     private val objectMapper: ObjectMapper,
     private val rabbitTemplate: RabbitTemplate
-) : ExternalScmResource {
+) : StreamExternalScmResource {
 
     companion object {
-        private val logger = LoggerFactory.getLogger(ExternalScmResourceImpl::class.java)
+        private val logger = LoggerFactory.getLogger(StreamExternalScmResourceImpl::class.java)
     }
 
     override fun webHookCodeGitCommit(token: String, eventType: String, event: String): Result<Boolean> {
