+++ src/backend/ci/core/stream/biz-stream/src/main/kotlin/com/tencent/devops/stream/resources/external/StreamExternalStreamResourceImpl.kt
@@ -25,22 +25,27 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.misc.config
+package com.tencent.devops.stream.resources.external
 
-import org.springframework.context.annotation.Configuration
-import org.springframework.scheduling.annotation.SchedulingConfigurer
-import org.springframework.scheduling.config.ScheduledTaskRegistrar
-import java.util.concurrent.ScheduledThreadPoolExecutor
-import java.util.concurrent.ThreadPoolExecutor
-import java.util.concurrent.TimeUnit
+import com.tencent.devops.common.web.RestResource
+import com.tencent.devops.stream.api.external.StreamExternalStreamResource
+import com.tencent.devops.stream.service.StreamPipelineBadgeService
 
-@Configuration
-class ScheduleConfig : SchedulingConfigurer {
-    override fun configureTasks(taskRegistrar: ScheduledTaskRegistrar) {
-        val scheduler = ScheduledThreadPoolExecutor(20)
-        scheduler.maximumPoolSize = 500
-        scheduler.setKeepAliveTime(60, TimeUnit.SECONDS)
-        scheduler.rejectedExecutionHandler = ThreadPoolExecutor.CallerRunsPolicy()
-        taskRegistrar.setScheduler(scheduler)
+@RestResource
+class StreamExternalStreamResourceImpl(
+    private val streamPipelineBadgeService: StreamPipelineBadgeService
+) : StreamExternalStreamResource {
+    override fun getPipelineBadge(
+        gitProjectId: Long,
+        filePath: String,
+        branch: String?,
+        objectKind: String?
+    ): String {
+        return streamPipelineBadgeService.get(
+            gitProjectId = gitProjectId,
+            filePath = filePath,
+            branch = branch,
+            objectKind = objectKind
+        )
     }
 }
